requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
fpdf2>=2.7.0
python-multipart>=0.0.6
aiosmtplib>=3.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
